ENH: move and apply formatSetQUeryParams to utils folder
